Add a CI badge and new reference to README, update packages version, add tests documentation
The project has been updated with various enhancements. The README files now feature a CI badge and a new reference. The versions of all referenced packages have been upgraded to the latest, boosting performance and security. A detailed explanation on testing, along with conventions has been provided in a README file in the test directory. An OpenTelemetry package has been added to the project.